+++ api/responses/tyre_impression_responses.py
@@ -1,3 +1,4 @@
+from api.responses.tyre_impression_processing_responses import tyre_impression_processing_response
 from database.models.tyre_impression import TyreImpression
 
 
@@ -7,4 +8,6 @@ def tyre_impression_response(tyre_impression: TyreImpression):
         "uuid": tyre_impression.uuid,
         "status": tyre_impression.status.value,
         "created_at": tyre_impression.created_at.isoformat() if tyre_impression.created_at else None,
+
+        "processing": tyre_impression_processing_response(tyre_impression.processing) if tyre_impression.processing else None,
     }