+++ tests/api/routes/test_tyre_impression_routes.py
@@ -2,12 +2,16 @@ import http
 from unittest.mock import patch
 from werkzeug.datastructures import FileStorage
 from database.models.tyre_impression import TyreImpression
+from tests.helpers.factories.file_factory import FileFactory
+from database.models.data_types.files import FileModel, FileType
 from services.tyre_impression_service import TyreImpressionService
 from domain.exceptions import InvalidFileTypeError, DatabaseError, FileSaveError
 from tests.helpers.factories.tyre_impression_factory import TyreImpressionFactory
 from database.models.data_types.tyre_impression_status import TyreImpressionStatus
+from tests.helpers.factories.tyre_impression_processing_factory import TyreImpressionProcessingFactory
 from tests.helpers.assertions import assert_paginated_response, assert_tyre_impression_response, \
-    assert_tyre_impression_not_in_response, assert_error_response
+    assert_tyre_impression_not_in_response, assert_error_response, assert_is_not_tyre_impression, \
+    assert_tyre_impression_processing_response, assert_file_response
 
 
 def test_get_all_empty_response(client):
@@ -79,6 +83,44 @@ def test_get_all_pagination_get_page_2(client):
     # Ensure other models were not returned
     assert_tyre_impression_not_in_response(data, tyre_impression_1)
     assert_tyre_impression_not_in_response(data, tyre_impression_3)
+
+
+def test_get_by_id_invalid_id(client):
+    response = client.get(f"/tyre-impressions/1000")
+
+    # Ensure correct status code and error message were returned
+    assert_error_response(response, http.HTTPStatus.NOT_FOUND, "Tyre impression with id 1000 not found")
+
+
+def test_get_by_id(client, database_session):
+    tyre_impression_1 = TyreImpressionFactory.create()
+    tyre_impression_2 = TyreImpressionFactory.create()
+
+    tyre_impression_processing = TyreImpressionProcessingFactory.create(tyre_impression_1.id)
+
+    original_file = FileFactory.create(FileModel.tyre_impression, tyre_impression_processing.id, file_type=FileType.original)
+    normalised_file = FileFactory.create(FileModel.tyre_impression, tyre_impression_processing.id, file_type=FileType.normalised)
+    enhanced_file = FileFactory.create(FileModel.tyre_impression, tyre_impression_processing.id, file_type=FileType.enhanced)
+
+    response = client.get(f"/tyre-impressions/{tyre_impression_1.id}")
+
+    assert http.HTTPStatus.OK == response.status_code
+    data = response.get_json()
+    assert_tyre_impression_response(data, tyre_impression_1)
+    assert_is_not_tyre_impression(data, tyre_impression_2)
+
+    # Ensure processing object was joined
+    assert data["processing"] is not None
+    assert_tyre_impression_processing_response(data["processing"], tyre_impression_processing)
+
+    # Ensure all files were joined to the processing object
+    assert data["processing"]["files"] is not None
+    assert FileType.original.value in data["processing"]["files"]
+    assert_file_response(data["processing"]["files"][FileType.original.value], original_file)
+    assert FileType.normalised.value in data["processing"]["files"]
+    assert_file_response(data["processing"]["files"][FileType.normalised.value], normalised_file)
+    assert FileType.enhanced.value in data["processing"]["files"]
+    assert_file_response(data["processing"]["files"][FileType.enhanced.value], enhanced_file)
 
 
 def test_upload_no_file(client):